refactored code into their own files
