+++ src/constants.ts
@@ -1,0 +1,6 @@
+export const TEXT_PADDING_LR = 60;
+export const TEXT_PADDING_UD = 33;
+export const IO_PADDING_LR = 10;
+export const IO_PADDING_UD = 20;
+export const IO_TEXT_PADDING_LR = 15;
+export const IO_TEXT_PADDING_UD = 10;
