+++ src/editorElement.ts
@@ -1,0 +1,7 @@
+export class EditorElement {
+    props: any;
+  
+    constructor(props) {
+      this.props = props;
+    }
+  }
